Fix Issue #55
hot reload: resolved segfault in refresh signal
libout: duplicates in output

As part of the hot reload, the refresh signal did not ensure that the socket was open before transmitting data over the line. To resolve this, two new methods were added "IsActive" & "IsActiveBundle". These help ensure that the socket is first open and primed before usage. 

In addition to this segfault, the component packer would incorrectly determine the change request type, due to the socket not being established first, and part of the match for preexisting pages required a valid signal. 

Fixes #54  ; Fixes #55