#33 recompute bundle upon redirection of page during hot reload (#61)
added recomputation of bundle support during hot reload in dev command

A channel was made to trigger the refresh event within the application. The channel will then recompute the bundle (provided the bundle key) and send a refresh signal back to the browser. Additionally, a check was put in place to ensure that the bundle keys do not get recomputed during excessive refreshing of the browser. 

Fixes #33