vanilla JS support & micro-frontend #11  (#71)
webwrap: vanilla JS support i.e support of the extension ".js" in the pages directory

Added support for vanilla javascript pages in the pages directory. In addition to this, a new operation exists on the libout orbit.Request called "RenderPages" that allows multiple pages to be rendered in a micro-frontend fashion. 

Fixes #11