fix hotreload unable to detect bundle for micro frontends (#77) (#79)
hotreload: fixed support for micro frontends

During adding support for micro frontends, a change was made to allow support for multiple bundle keys within the hot reloader which ultimately broke hot reload. So to resolve this issue, support was added for diffing bundle keys to determine which of the keys should be bundled. 

Fixes #77