dev cmd: lru caching support for hot reload redirection #62 (#88)
dev cmd: LRU caching support for hot reload redirection

Added caching support for bundle redirection from the hot reloader. Previously, within dev mode, if a page redirection was detected it would recompute the page always. This was done to ensure that the page was not a stale copy when viewed, but we only really wanted this to happen once for every page upon first recomputation. To resolve this issue a stack-based cache was applied for the last 10 recently computed pages, invalidating using LRU. 

Fixes #62