support for static deployment command (#93)
support for static deployment command

Added basic support for a deployment command to output static files. The command can be accessed with the `deploy` keyword within the CLI and currently accepts all the same flags as the build command, in addition to a "staticout" flag to specify the directory to output the static content. 

Fixes #93