fix go generation variable reordering issues  (#124)
Resolved issues around orbit go code generation causing changes regarding the ordering of various variable elements. This was solved by ordering the page names before serialization as well as resolving an issue that caused the id of react div frames to be recomputed upon every change request. 

This fix also resolved issue #123, which caused a single div to be rendered instead of the necessary divs during the usage of micro frontends.  

Fixed #123 
Fixed #122