dev cmd: improve reoccurring startup duration  (#127)
Dev cmd improvements that increase startup speeds for recurring reboots of the application. These fixes are measured in a > 100x speedup in boot times where the application has already initialized the required bundle data.

In addition, a flag was created for the dev cmd, that terminates the application after startup has been completed. This command should be used in the future to create a test that can validate regressions in the startup duration. 

Fix #112