Improved lighthouse scores (#142)
Benchmarked the example application with lighthouse, identified key heuristics that required improvement, and improved the scores for the autogenerated section of the go output. 

This also includes a redesign of the e2e test suite around modular tests, including the addition of a lighthouse CLI wrapper that ensures optimal performance for the following heuristics.

- first-contentful-paint
- speed-index
- largest-contentful-paint
- server-response-time

Additionally, a bug was resolved regarding the autogen ENV not being propagated correctly. 

Fix #145
Fix #140