Feat/#173: Added SPA Support (#175)
Added support for React SPA (single page applications). This change introduces two new flags `spa_entry_path` and `spa_out_dir` that when set will build an SPA.

SPA is currently only supported for the build command, not currently supported for the dev command. 

This change also introduces a change to all of the command flags, reflected in the documentation during the next release build.